refactoring userManagement util https://github.com/dorbel-tech/dorbel-app/issues/949
