+++ src/utils/user/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  helpers: require('helpers'),
+  management: require('management'),
+  permissions: require('permissions'),
+};
